Resolve complex utils/files merge.
