+++ utils/bids_files.js
@@ -3,7 +3,7 @@
  * eg. corresponding files
  */
 // dependencies -------------------------------------------------------------------
-var files = require('./files')
+const dataExtRE = require('./data_extensions.js')
 
 // public API ---------------------------------------------------------------------
 const bidsFileUtils = {
@@ -64,7 +64,7 @@ function verifyDatafileMatch(sidecarPath, noExt, matchFile) {
   let megDs = false
   // Make sure it's not the data dictionary itself
   const isSelf = matchFile.relativePath === sidecarPath
-  if (!isSelf && files.dataExtRE().test(matchFile.name)) {
+  if (!isSelf && dataExtRE.test(matchFile.name)) {
     match = true
   }
   // MEG datafiles may be a folder, therefore not contained in fileList, will need to look in paths
