+++ utils/files/readFile.js
@@ -14,42 +14,34 @@ const isNode = typeof window === 'undefined'
  * In node the file should be a path to a file.
  *
  */
-function readFile(file, callback) {
-  if (isNode) {
-    readNodeFile(file, callback)
-  } else {
-    readBrowserFile(file, callback)
-  }
-}
-
-function readNodeFile(file, callback) {
-  testFile(file, function(issue) {
-    if (issue) {
-      process.nextTick(function() {
-        callback(issue, null)
+function readFile(file) {
+  return new Promise((resolve, reject) => {
+    if (fs) {
+      testFile(file, function(issue) {
+        if (issue) {
+          process.nextTick(function() {
+            return reject(issue)
+          })
+        }
+        fs.readFile(file.path, 'utf8', function(err, data) {
+          process.nextTick(function() {
+            return resolve(data)
+          })
+        })
       })
-      return
-    }
-    fs.readFile(file.path, 'utf8', function(err, data) {
-      process.nextTick(function() {
-        callback(null, data)
-      })
-    })
-  })
-}
-
-function readBrowserFile(file, callback) {
-  var reader = new FileReader()
-  reader.onloadend = function(e) {
-    if (e.target.readyState == FileReader.DONE) {
-      if (!e.target.result) {
-        callback(new Issue({ code: 44, file: file }), null)
-        return
+    } else {
+      var reader = new FileReader()
+      reader.onloadend = function(e) {
+        if (e.target.readyState == FileReader.DONE) {
+          if (!e.target.result) {
+            return reject(new Issue({ code: 44, file: file }))
+          }
+          return resolve(e.target.result)
+        }
       }
-      callback(null, e.target.result)
+      reader.readAsBinaryString(file)
     }
-  }
-  reader.readAsBinaryString(file)
+  })
 }
 
 module.exports = readFile
